requirements.txt
google-genai==1.5.0
gradio>=4.0.0
pydantic==2.10.6
Pillow>=10.0.0
python-multipart>=0.0.6
numpy>=1.24.0
typing-extensions>=4.5.0